ai-tools/silicon-studio-audit/backend/pyproject.toml
[project]
name = "perimeter-backend"
version = "0.1.0"
description = "Local-first LLM fine-tuning backend for Perimeter.ai"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "mlx-lm>=0.1.0",
    "pandas>=2.2.0",
    "presidio-analyzer>=2.2.351",
    "presidio-anonymizer>=2.2.351",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2.0",
    "psutil>=5.9.8"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.13.0"
]

[tool.setuptools]
packages = ["app"]
